Sitemap XML generator
- refactoring & tests
